Separate the batch file into init and run
PYQUM_init.bat :
for first time install the server, it will create a new file "path.cfg" beside the folder of repository to store the path of conda and database.
SERVER_RUN.bat :
for start the server
IMPORTANT!! The all function for calling database in python code still as same as before.