Removed transpilation to recover decay
The issue was that gates appended to the circuits (the ones where the noise model was supposed to bind the depolarizing noise) were decomposed further through transpilation (optimized decomposition) making the whole binding process uneffective. Transpilation has been removed and linear XEB (Google post processing) works fine. However log estimation still fails.

Also added new metrics such as fidelities_list and purities_list. Thos enable the checking of the output fidelity of the simulated density matrix (with noise) and the ideal statevector. We can see that the fidelity decays to 0.5 which makes sense for a completely mixed state. The density matrices are readily accessible as well.